adjust the file structure and add file utils.py to store tool functions or tool classes
